Create Readme to keep track of what needs to be/what is done.
I might change that to a devlog in a web platform when at least the basic structure is done and everything else are features.
The web app will be responsible of account and character creation and also basic information about the game and how it works.
This game will not have a Wiki or a tutorial of sorts for everything. This is meant to be a hard to explore game.